refactor: move genres description data to consts
